Enable MNNVLAllreduce automatically when MNNVL is available

+++ tensorrt_llm/_torch/modules/fused_moe/interface.py
@@ -86,7 +86,8 @@ class MoE(nn.Module):
         self.intermediate_size_per_partition = intermediate_size // self.tp_size
 
         self.all_reduce = AllReduce(mapping=self.mapping,
-                                    strategy=model_config.allreduce_strategy)
+                                    strategy=model_config.allreduce_strategy,
+                                    dtype=self.dtype)
 
     @abstractmethod
     def create_weights(self):
